+++ src/blob.rs
@@ -16,13 +16,13 @@ use std::{
     u64,
 };
 
-/// Maximum allowed size for a serialised ImmutableData to grow to.
-pub const MAX_IMMUTABLE_DATA_SIZE_IN_BYTES: u64 = 1024 * 1024 + 10 * 1024;
+/// Maximum allowed size for a serialised Blob to grow to.
+pub const MAX_BLOB_SIZE_IN_BYTES: u64 = 1024 * 1024 + 10 * 1024;
 
-/// Unpublished ImmutableData: an immutable chunk of data which can be deleted. Can only be fetched
+/// Unpublished Blob: an immutable chunk of data which can be deleted. Can only be fetched
 /// by the listed owner.
 #[derive(Hash, Eq, PartialEq, PartialOrd, Ord, Clone)]
-pub struct UnpubData {
+pub struct PrivateData {
     /// Network address. Omitted when serialising and calculated from the `value` and `owner` when
     /// deserialising.
     address: Address,
@@ -33,8 +33,8 @@ pub struct UnpubData {
     owner: PublicKey,
 }
 
-impl UnpubData {
-    /// Creates a new instance of `UnpubData`.
+impl PrivateData {
+    /// Creates a new instance of `PrivateData`.
     pub fn new(value: Vec<u8>, owner: PublicKey) -> Self {
         let hash_of_value = tiny_keccak::sha3_256(&value);
         let serialised_contents = utils::serialise(&(hash_of_value, &owner));
@@ -79,33 +79,33 @@ impl UnpubData {
 
     /// Returns `true` if the size is valid.
     pub fn validate_size(&self) -> bool {
-        self.serialised_size() <= MAX_IMMUTABLE_DATA_SIZE_IN_BYTES
+        self.serialised_size() <= MAX_BLOB_SIZE_IN_BYTES
     }
 }
 
-impl Serialize for UnpubData {
+impl Serialize for PrivateData {
     fn serialize<S: Serializer>(&self, serialiser: S) -> Result<S::Ok, S::Error> {
         (&self.value, &self.owner).serialize(serialiser)
     }
 }
 
-impl<'de> Deserialize<'de> for UnpubData {
+impl<'de> Deserialize<'de> for PrivateData {
     fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
         let (value, owner): (Vec<u8>, PublicKey) = Deserialize::deserialize(deserializer)?;
-        Ok(UnpubData::new(value, owner))
+        Ok(PrivateData::new(value, owner))
     }
 }
 
-impl Debug for UnpubData {
+impl Debug for PrivateData {
     fn fmt(&self, formatter: &mut Formatter) -> fmt::Result {
         // TODO: Output owners?
-        write!(formatter, "UnpubImmutableData {:?}", self.name())
+        write!(formatter, "UnpubBlob {:?}", self.name())
     }
 }
 
-/// Published ImmutableData: an immutable chunk of data which cannot be deleted.
+/// Published Blob: an immutable chunk of data which cannot be deleted.
 #[derive(Hash, Clone, Eq, PartialEq, Ord, PartialOrd)]
-pub struct PubData {
+pub struct PublicData {
     /// Network address. Omitted when serialising and calculated from the `value` when
     /// deserialising.
     address: Address,
@@ -113,8 +113,8 @@ pub struct PubData {
     value: Vec<u8>,
 }
 
-impl PubData {
-    /// Creates a new instance of `ImmutableData`.
+impl PublicData {
+    /// Creates a new instance of `Blob`.
     pub fn new(value: Vec<u8>) -> Self {
         Self {
             address: Address::Pub(XorName(tiny_keccak::sha3_256(&value))),
@@ -149,30 +149,30 @@ impl PubData {
 
     /// Returns true if the size is valid.
     pub fn validate_size(&self) -> bool {
-        self.serialised_size() <= MAX_IMMUTABLE_DATA_SIZE_IN_BYTES
+        self.serialised_size() <= MAX_BLOB_SIZE_IN_BYTES
     }
 }
 
-impl Serialize for PubData {
+impl Serialize for PublicData {
     fn serialize<S: Serializer>(&self, serialiser: S) -> Result<S::Ok, S::Error> {
         self.value.serialize(serialiser)
     }
 }
 
-impl<'de> Deserialize<'de> for PubData {
+impl<'de> Deserialize<'de> for PublicData {
     fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
         let value: Vec<u8> = Deserialize::deserialize(deserializer)?;
-        Ok(PubData::new(value))
+        Ok(PublicData::new(value))
     }
 }
 
-impl Debug for PubData {
+impl Debug for PublicData {
     fn fmt(&self, formatter: &mut Formatter) -> fmt::Result {
-        write!(formatter, "PubImmutableData {:?}", self.name())
+        write!(formatter, "PubBlob {:?}", self.name())
     }
 }
 
-/// Kind of an ImmutableData.
+/// Kind of an Blob.
 #[derive(Clone, Copy, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize, Debug)]
 pub enum Kind {
     /// Unpublished.
@@ -202,7 +202,7 @@ impl Kind {
     }
 }
 
-/// Address of an ImmutableData.
+/// Address of an Blob.
 #[derive(Clone, Copy, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize, Debug)]
 pub enum Address {
     /// Unpublished namespace.
@@ -256,13 +256,13 @@ impl Address {
     }
 }
 
-/// Object storing an ImmutableData variant.
+/// Object storing an Blob variant.
 #[derive(Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize, Debug)]
 pub enum Data {
-    /// Unpublished ImmutableData.
-    Unpub(UnpubData),
-    /// Published ImmutableData.
-    Pub(PubData),
+    /// Unpublished Blob.
+    Unpub(PrivateData),
+    /// Published Blob.
+    Pub(PublicData),
 }
 
 impl Data {
@@ -319,21 +319,21 @@ impl Data {
     }
 }
 
-impl From<UnpubData> for Data {
-    fn from(data: UnpubData) -> Self {
+impl From<PrivateData> for Data {
+    fn from(data: PrivateData) -> Self {
         Data::Unpub(data)
     }
 }
 
-impl From<PubData> for Data {
-    fn from(data: PubData) -> Self {
+impl From<PublicData> for Data {
+    fn from(data: PublicData) -> Self {
         Data::Pub(data)
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::{utils, Address, PubData, PublicKey, UnpubData, XorName};
+    use super::{utils, Address, PrivateData, PublicData, PublicKey, XorName};
     use bincode::deserialize as deserialise;
     use hex::encode;
     use rand::{self, Rng, SeedableRng};
@@ -350,10 +350,10 @@ mod tests {
         let owner1 = PublicKey::Bls(SecretKey::random().public_key());
         let owner2 = PublicKey::Bls(SecretKey::random().public_key());
 
-        let idata1 = UnpubData::new(data1.clone(), owner1);
-        let idata2 = UnpubData::new(data1, owner2);
-        let idata3 = UnpubData::new(data2.clone(), owner1);
-        let idata3_clone = UnpubData::new(data2, owner1);
+        let idata1 = PrivateData::new(data1.clone(), owner1);
+        let idata2 = PrivateData::new(data1, owner2);
+        let idata3 = PrivateData::new(data2.clone(), owner1);
+        let idata3_clone = PrivateData::new(data2, owner1);
 
         assert_eq!(idata3, idata3_clone);
 
@@ -365,11 +365,11 @@ mod tests {
     #[test]
     fn deterministic_test() {
         let value = "immutable data value".to_owned().into_bytes();
-        let immutable_data = PubData::new(value);
-        let immutable_data_name = encode(immutable_data.name().0.as_ref());
+        let blob = PublicData::new(value);
+        let blob_name = encode(blob.name().0.as_ref());
         let expected_name = "fac2869677ee06277633c37ac7e8e5c655f3d652f707c7a79fab930d584a3016";
 
-        assert_eq!(&expected_name, &immutable_data_name);
+        assert_eq!(&expected_name, &blob_name);
     }
 
     #[test]
@@ -377,10 +377,10 @@ mod tests {
         let mut rng = get_rng();
         let len = rng.gen_range(1, 10_000);
         let value = iter::repeat_with(|| rng.gen()).take(len).collect();
-        let immutable_data = PubData::new(value);
-        let serialised = utils::serialise(&immutable_data);
+        let blob = PublicData::new(value);
+        let serialised = utils::serialise(&blob);
         let parsed = unwrap!(deserialise(&serialised));
-        assert_eq!(immutable_data, parsed);
+        assert_eq!(blob, parsed);
     }
 
     fn get_rng() -> XorShiftRng {