+++ CHANGELOG.md
@@ -93,7 +93,7 @@ All notable changes to this project will be documented in this file. See [standa
 - Added the identities (public and private) for clients and apps.
 - Added the AppendOnlyData types.
 - Added the unpublished MutableData type.
-- Added the unpublished and published ImmutableData types.
+- Added the unpublished and published Blob types.
 - Added the Request and Response RPCs.
 
 ## [0.1.0]